refactor: change file validation to happen explicitly
Previously, it happened at config resolution time, which was awkward because
some commands didn't care about the list of files.

This also sets us up to make the set of files more flexible.
